Assembled the PC's model
I put the player character's model into unity and assembled her (I
applied the textures to her body and her armor and the assigned the
various pieces of her armor to the corresponding parts of her armature).
She also has an animation controller but it doesn't really have her
animations put in other than a couple. I'll leave that to player
control.

I do NOT have cloth working for her...