Menu pictures
As always this is more of a rough draft than the final product (Unless everyone loves it).  I am hoping to make it controlled by the keyboard instead of the mouse and add a quit button (Or something.) but the way I have it setup right now I think only one button would work.