Imported Models, gave player sword
I've imported all of my models as .fbx files and applied textures or
materials to most of them. I made the knife red-orange, but feel free to
change that.

Test--Slash, Sword--Heavy, and Sword--Light animations are doing
something weird, so I think we should just use the Halberd animations
because they work.