It's not bug free yet. But the wasp chases you
It also makes a noise when it is chasing you. (which needs a little editing to loop correctly.)  It's not bug free yet. If you bump it from below it goes flying into the sky. Not sure why. It doesn't damage the player yet.
You also need to add the player from the scene to the prefab for it to work right.