Level Music!
I like the harder edge of the final level music, but it clashes with the new win music.  So maybe I'll change them around again.  Let me know what you think.  I picked an upbeat friendlier tune for the first level and the old Win music is now the 4th level because it seemed more Autumn-y. But that seemed like an active level, so maybe that doesn't fit exactly either.