wasp OnCollisionEnter  player
Added a public variable of "strength" for how much damage the wasp does to the player. Apparently I needed to call an attack on the player when they did an OnCollisionEnter.  Or at least, this is what fixed the wasp not hurting the player when it was not a trigger.