Forgot to add the sound script
Made a slight change to Arrow and Knife Manager scripts so they be sure to send damage before destroying themselves.  I also found a script bit in one of the player managers that might explain why the wasp would hurt the player when she killed it.  I changed it to a public int and kept the default value of 20. (which seemed a bit high to me.)