Camera pull on all levels
And a few other changes.  Wasps are a little faster but don't engage from quite so far away.  Adjusted volume of music and sound effects on all levels. Hopefully that sounds better. Changed the attack distance of the turrets on the final level. (So that they aren't attacking while the camera pulls back.) Changed wasp sound, and it only plays once.  (I think this should work on everyone's machine. I'm pretty sure I uploaded the sound a long while back.)